Move CM.from_langchain logic to service.utils
